@@ -1,7 +1,20 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import 'antd/dist/antd.css';
-import './index.css';
+import App from "./App";
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { App as AntApp, ConfigProvider } from "antd";
 
-ReactDOM.render(<App />, document.getElementById('root'));
+import "./index.css";
+
+ReactDOM.createRoot(document.getElementById("root")).render(
+  <ConfigProvider
+    theme={{
+      hashed: false,
+      components: { Button: { fontWeight: 500 } },
+      token: { controlHeight: 36, colorPrimary: "#2196f3", fontFamily: "Inter" },
+    }}
+  >
+    <AntApp>
+      <App />
+    </AntApp>
+  </ConfigProvider>
+);
